refactor: move 6 facade files from src/utils/ to src/facades/
Move board-facade, file-tree-facade, skills-facade, tab-facade,
tab-manager-api, and terminal-panel-facade out of src/utils/ into a
dedicated src/facades/ directory, resolving the layer-crossing where
utils/ modules imported directly from services/.

All consumer imports updated (components + utils helpers).
Build and tests pass (27 files, 400 tests).

Closes #520
