+++ src/facades/tab-manager-api.js
