+++ src/facades/terminal-panel-facade.js
